Moving existing bounding boxes
